Fix call of server handle in Gevents Hub context
Before that patch, request handler of the server was executed in
Gevents Hub context. Which is wrong, because inside it it is
impossible to use Gevent blocking API, and servers handle uses it.

This patch fixed it - instead of overriding `do_handle` method, which is not
supposed to be overriden, server just uses public interface of StreamServer
and passes handle to it's init.